requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.0.20
langchain==0.1.0
langchain-mistralai==0.0.1
langchain-community==0.0.10
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
yfinance==0.2.28
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
beautifulsoup4==4.12.2
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
aiofiles==23.2.1
newsapi-python==0.2.6
plotly==5.17.0
streamlit==1.28.1
